+++ lib/action_controller/serialization.rb
@@ -29,6 +29,11 @@ module ActionController
 
     include ActionController::Renderers
 
+    class << self
+      attr_accessor :enabled
+    end
+    self.enabled = true
+
     included do
       class_attribute :_serialization_scope
       self._serialization_scope = :current_user
@@ -52,6 +57,18 @@ module ActionController
 
     private
 
+    def namespace_for_serializer
+      @namespace_for_serializer ||= self.class.parent unless self.class.parent == Object
+    end
+
+    def default_serializer(resource)
+      options = {}.tap do |o|
+        o[:namespace] = namespace_for_serializer if namespace_for_serializer
+      end
+
+      ActiveModel::Serializer.serializer_for(resource, options)
+    end
+
     def default_serializer_options
       {}
     end
@@ -63,10 +80,15 @@ module ActionController
 
     def build_json_serializer(resource, options = {})
       options = default_serializer_options.merge(options)
+      @namespace_for_serializer = options.fetch(:namespace, nil)
 
-      if serializer = options.fetch(:serializer, ActiveModel::Serializer.serializer_for(resource))
+      if serializer = options.fetch(:serializer, default_serializer(resource))
         options[:scope] = serialization_scope unless options.has_key?(:scope)
-        options[:resource_name] = controller_name if resource.respond_to?(:to_ary)
+
+        if resource.respond_to?(:each)
+          options[:resource_name] = controller_name 
+          options[:namespace] = namespace_for_serializer if namespace_for_serializer
+        end
 
         serializer.new(resource, options)
       end