+++ lib/active_model/serializable.rb
@@ -1,11 +1,25 @@
+require 'active_model/serializable/utils'
+
 module ActiveModel
   module Serializable
+    def self.included(base)
+      base.extend Utils
+    end
+
     def as_json(options={})
-      if root = options.fetch(:root, json_key)
-        hash = { root => serializable_object }
-        hash.merge!(serializable_data)
-        hash
-      else
+      instrument('!serialize') do
+        if root = options.fetch(:root, json_key)
+          hash = { root => serializable_object }
+          hash.merge!(serializable_data)
+          hash
+        else
+          serializable_object
+        end
+      end
+    end
+
+    def serializable_object_with_notification
+      instrument('!serialize') do
         serializable_object
       end
     end
@@ -18,8 +32,31 @@ module ActiveModel
       end
     end
 
+    def namespace
+      get_namespace && Utils._const_get(get_namespace)
+    end
+
     def embedded_in_root_associations
       {}
+    end
+
+    private
+
+    def get_namespace
+      modules = self.class.name.split('::')
+      modules[0..-2].join('::') if modules.size > 1
+    end
+
+    def instrument(action, &block)
+      payload = instrumentation_keys.inject({ serializer: self.class.name }) do |payload, key|
+        payload[:payload] = self.instance_variable_get(:"@#{key}")
+        payload
+      end
+      ActiveSupport::Notifications.instrument("#{action}.active_model_serializers", payload, &block)
+    end
+
+    def instrumentation_keys
+      [:object, :scope, :root, :meta_key, :meta, :wrap_in_array, :only, :except, :key_format]
     end
   end
 end