+++ test/fixtures/poro.rb
@@ -1,11 +1,13 @@
 class Model
-  def initialize(hash={})
+  def initialize(hash = {})
     @attributes = hash
   end
 
   def read_attribute_for_serialization(name)
     if name == :id || name == 'id'
       object_id
+    elsif respond_to?(name)
+      send name
     else
       @attributes[name]
     end
@@ -22,6 +24,12 @@ class User < Model
   end
 end
 
+class UserInfo < Model
+  def user
+    @user ||= User.new(name: 'N1', email: 'E1')
+  end
+end
+
 class Admin < Model
   def profile
     @profile ||= Profile.new(name: 'N2', description: 'D2')
@@ -29,6 +37,13 @@ class Admin < Model
 end
 
 class Profile < Model
+end
+
+class Category < Model
+  def posts
+    @posts ||= [Post.new(title: 'T1', body: 'B1'),
+                Post.new(title: 'T2', body: 'B2')]
+  end
 end
 
 class Post < Model
@@ -41,6 +56,28 @@ end
 class Comment < Model
 end
 
+class WebLog < Model
+end
+
+class Interview < Model
+  def attachment
+    @attachment ||= Image.new(url: 'U1')
+  end
+end
+
+class Mail < Model
+  def attachments
+    @attachments ||= [Image.new(url: 'U1'),
+                      Video.new(html: 'H1')]
+  end
+end
+
+class Image < Model
+end
+
+class Video < Model
+end
+
 ###
 ## Serializers
 ###
@@ -50,6 +87,7 @@ class UserSerializer < ActiveModel::Serializer
   has_one :profile
 end
 
+<<<<<<< HEAD
 class AdminSerializer < ActiveModel::Serializer
   attributes :name, :language
   flattened_attributes profile_name: [:profile, :name]
@@ -57,6 +95,10 @@ class AdminSerializer < ActiveModel::Serializer
   def language
     'en'
   end
+=======
+class UserInfoSerializer < ActiveModel::Serializer
+  has_one :user
+>>>>>>> rails-api/0-9-stable
 end
 
 class ProfileSerializer < ActiveModel::Serializer
@@ -68,6 +110,16 @@ class ProfileSerializer < ActiveModel::Serializer
   attributes :name, :description
 end
 
+class DifferentProfileSerializer < ActiveModel::Serializer
+  attributes :name
+end
+
+class CategorySerializer < ActiveModel::Serializer
+  attributes :name
+
+  has_many :posts
+end
+
 class PostSerializer < ActiveModel::Serializer
   attributes :title, :body
 
@@ -76,4 +128,39 @@ end
 
 class CommentSerializer < ActiveModel::Serializer
   attributes :content
+end
+
+class WebLogSerializer < ActiveModel::Serializer
+  attributes :name, :display_name
+end
+
+class WebLogLowerCamelSerializer < WebLogSerializer
+  format_keys :lower_camel
+end
+
+class InterviewSerializer < ActiveModel::Serializer
+  attributes :text
+
+  has_one :attachment, polymorphic: true
+end
+
+class MailSerializer < ActiveModel::Serializer
+  attributes :body
+
+  has_many :attachments, polymorphic: true
+end
+
+class ImageSerializer < ActiveModel::Serializer
+  attributes :url
+end
+
+class VideoSerializer < ActiveModel::Serializer
+  attributes :html
+end
+
+class ShortProfileSerializer < ::ProfileSerializer; end
+
+module TestNamespace
+  class ProfileSerializer < ::ProfileSerializer; end
+  class UserSerializer    < ::UserSerializer;    end
 end