+++ test/integration/action_controller/serialization_test.rb
@@ -194,6 +194,59 @@ module ActionController
       end
     end
 
+    class LowerCamelArraySerializerTest < ActionController::TestCase
+      class WebLogController < ActionController::Base
+        def render_array
+          render json: [WebLog.new({name: 'Name 1', display_name: 'Display Name 1'}), WebLog.new({name: 'Name 2', display_name: 'Display Name 2'})], each_serializer: WebLogLowerCamelSerializer
+        end
+      end
+
+      tests WebLogController
+
+      def test_render_array
+        get :render_array
+        assert_equal 'application/json', @response.content_type
+        assert_equal '{"webLog":[{"name":"Name 1","displayName":"Display Name 1"},{"name":"Name 2","displayName":"Display Name 2"}]}', @response.body
+      end
+    end
+
+    class LowerCamelWoRootSerializerTest < ActionController::TestCase
+      class WebLogController < ActionController::Base
+        def render_without_root
+          render json: WebLog.new({name: 'Name 1', display_name: 'Display Name 1'}), 
+                 root: false, 
+                 serializer: WebLogLowerCamelSerializer
+        end
+      end
+
+      tests WebLogController
+
+      def test_render_without_root
+        get :render_without_root
+        assert_equal 'application/json', @response.content_type
+        assert_equal '{"name":"Name 1","displayName":"Display Name 1"}', @response.body
+      end
+    end
+
+    class LowerCamelArrayWoRootSerializerTest < ActionController::TestCase
+      class WebLogController < ActionController::Base
+        def render_array_without_root
+          render json: [WebLog.new({name: 'Name 1', display_name: 'Display Name 1'}), 
+                        WebLog.new({name: 'Name 2', display_name: 'Display Name 2'})], 
+                 root: false, 
+                 each_serializer: WebLogLowerCamelSerializer
+        end
+      end
+
+      tests WebLogController
+
+      def test_render_array_without_root
+        get :render_array_without_root
+        assert_equal 'application/json', @response.content_type
+        assert_equal '[{"name":"Name 1","displayName":"Display Name 1"},{"name":"Name 2","displayName":"Display Name 2"}]', @response.body
+      end
+    end
+
     class ArrayEmbedingSerializerTest < ActionController::TestCase
       def setup
         super
@@ -228,6 +281,22 @@ module ActionController
         assert_equal 'application/json', @response.content_type
 
         assert_equal("{\"my\":[{\"name\":\"Name 1\",\"email\":\"mail@server.com\",\"profile_id\":#{@controller.user.profile.object_id}}],\"profiles\":[{\"name\":\"N1\",\"description\":\"D1\"}]}", @response.body)
+      end
+    end
+  
+    class ExplicitEachSerializerWithEnumarableObjectTest < ActionController::TestCase
+      class MyController < ActionController::Base
+        def render_array
+          render json: [Profile.new({ name: 'Name 1', description: 'Description 1', comments: 'Comments 1' })].to_enum, each_serializer: DifferentProfileSerializer
+        end
+      end
+
+      tests MyController
+
+      def test_render_array
+        get :render_array
+        assert_equal 'application/json', @response.content_type
+        assert_equal '{"my":[{"name":"Name 1"}]}', @response.body
       end
     end
   end