+++ test/unit/active_model/serializer/config_test.rb
@@ -67,6 +67,7 @@ module ActiveModel
         def test_apply_config_to_associations
           CONFIG.embed     = :ids
           CONFIG.embed_in_root = true
+          CONFIG.key_format = :lower_camel
 
           association = PostSerializer._associations[:comments]
           old_association = association.dup
@@ -76,6 +77,10 @@ module ActiveModel
           assert association.embed_ids?
           assert !association.embed_objects?
           assert association.embed_in_root
+          assert_equal :lower_camel, association.key_format
+          assert_equal 'post', PostSerializer.root_name
+          CONFIG.plural_default_root = true
+          assert_equal 'posts', PostSerializer.root_name
         ensure
           PostSerializer._associations[:comments] = old_association
           CONFIG.clear