+++ test/unit/active_model/serializer/has_many_test.rb
@@ -115,6 +115,23 @@ module ActiveModel
         }, @post_serializer.as_json)
       end
 
+      def test_associations_embedding_ids_including_objects_serialization_when_invoked_from_parent_serializer
+        @association.embed = :ids
+        @association.embed_in_root = true
+
+        category = Category.new(name: 'Name 1')
+        category.instance_variable_set(:@posts, [@post])
+        category_serializer = CategorySerializer.new(category)
+
+        assert_equal({
+          'category' => {
+            name: 'Name 1',
+            posts: [{ title: 'Title 1', body: 'Body 1', 'comment_ids' => @post.comments.map { |c| c.object_id } }]
+          },
+          comments: [{ content: 'C1' }, { content: 'C2' }]
+        }, category_serializer.as_json)
+      end
+
       def test_associations_embedding_nothing_including_objects_serialization_using_as_json
         @association.embed = nil
         @association.embed_in_root = true
@@ -166,6 +183,63 @@ module ActiveModel
 
         assert_equal({
           'post' => { title: 'Title 1', body: 'Body 1', comments: { my_content: ['fake'] } }
+        }, @post_serializer.as_json)
+      end
+
+      def test_associations_embedding_ids_including_objects_serialization_with_embed_in_root_key
+        @association.embed_in_root = true
+        @association.embed_in_root_key = :linked
+        @association.embed = :ids
+        assert_equal({
+          linked: {
+            comments: [
+              { content: 'C1' },
+              { content: 'C2' }
+            ]
+          },
+          'post' => {
+            title: 'Title 1', body: 'Body 1',
+            'comment_ids' => @post.comments.map(&:object_id)
+          }
+        }, @post_serializer.as_json)
+      end
+
+      def test_associations_embedding_ids_using_embed_namespace_including_object_serialization_with_embed_in_root_key
+        @association.embed_in_root = true
+        @association.embed_in_root_key = :linked
+        @association.embed = :ids
+        @association.embed_namespace = :links
+        @association.key = :comments
+        assert_equal({
+          linked: {
+            comments: [
+              { content: 'C1' },
+              { content: 'C2' }
+            ]
+          },
+          'post' => {
+            title: 'Title 1', body: 'Body 1',
+            links: {
+              comments: @post.comments.map(&:object_id)
+            }
+          }
+        }, @post_serializer.as_json)
+      end
+
+      def test_associations_embedding_objects_using_embed_namespace
+        @association.embed = :objects
+        @association.embed_namespace = :links
+
+        assert_equal({
+          'post' => {
+            title: 'Title 1', body: 'Body 1',
+            links: {
+              comments: [
+                { content: 'C1' },
+                { content: 'C2' }
+              ]
+            }
+          }
         }, @post_serializer.as_json)
       end
     end