+++ test/unit/active_model/serializer/has_one_test.rb
@@ -131,6 +131,20 @@ module ActiveModel
         }, @user_serializer.as_json)
       end
 
+      def test_associations_embedding_ids_including_objects_serialization_when_invoked_from_parent_serializer
+        @association.embed = :ids
+        @association.embed_in_root = true
+
+        user_info = UserInfo.new
+        user_info.instance_variable_set(:@user, @user)
+        user_info_serializer = UserInfoSerializer.new(user_info)
+
+        assert_equal({
+          'user_info' => { user: { name: 'Name 1', email: 'mail@server.com', 'profile_id' => @user.profile.object_id } },
+          'profiles' => [{ name: 'N1', description: 'D1' }]
+        }, user_info_serializer.as_json)
+      end
+
       def test_associations_embedding_ids_using_a_given_serializer
         @association.embed = :ids
         @association.embed_in_root = true
@@ -159,6 +173,47 @@ module ActiveModel
 
         assert_equal({
           'user' => { name: 'Name 1', email: 'mail@server.com', profile: { name: 'fake' } }
+        }, @user_serializer.as_json)
+      end
+
+      def test_associations_embedding_ids_using_embed_namespace
+        @association.embed_namespace = :links
+        @association.embed = :ids
+        @association.key = :profile
+        assert_equal({
+          'user' => {
+            name: 'Name 1', email: 'mail@server.com',
+            links: {
+              profile: @user.profile.object_id
+            }
+          }
+        }, @user_serializer.as_json)
+      end
+
+      def test_asociations_embedding_objects_using_embed_namespace
+        @association.embed_namespace = :links
+        @association.embed = :objects
+        assert_equal({
+          'user' => {
+            name: 'Name 1', email: 'mail@server.com',
+            links: {
+              profile: { name: 'N1', description: 'D1' }
+            }
+          }
+        }, @user_serializer.as_json)
+      end
+
+      def test_associations_embedding_ids_using_embed_namespace_and_embed_in_root_key
+        @association.embed_in_root = true
+        @association.embed_in_root_key = :linked
+        @association.embed = :ids
+        assert_equal({
+          'user' => {
+            name: 'Name 1', email: 'mail@server.com', 'profile_id' => @user.profile.object_id
+          },
+          linked: {
+            'profiles' => [ { name: 'N1', description: 'D1' } ]
+          }
         }, @user_serializer.as_json)
       end
     end